+++ resnap/__init__.py
@@ -3,7 +3,7 @@
 from .decorators import async_resnap, resnap
 from .exceptions import ResnapError
 from .factory import set_resnap_service
-from .helpers.context import add_metadata, add_metadatas
+from .helpers.context import add_metadata, add_multiple_metadata
 from .services.service import ResnapService
 from .version import VERSION
 
@@ -16,9 +16,9 @@ __all__ = (
     "ResnapError",
     # factory
     "set_resnap_service",
-    # metadatas
+    # metadata
     "add_metadata",
-    "add_metadatas",
+    "add_multiple_metadata",
     # services
     "ResnapService",
     # version