+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2025-06-19
+### Added
+- Add mypy and pyright support
+
+### Changed
+- Change hash logic
+- Rename add_metadatas -> add_multiple_metadata
+- Change CI with #no_test tag in commit message
+
+### Fixed
+- Fix config check secret filename if not enabled
+- Fix some typing
+
 ## [0.1.1] - 2025-05-22
 ### Changed
 - Fix documentation with Github URL